TT-001: remove classes related to operations

+++ src/main/java/az/technical/task/technicaltask/repository/TransferRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TransferRepository extends JpaRepository<TransferEntity, Long> {
+public interface TransferRepository extends JpaRepository<TransferEntity, Integer> {
 
     List<TransferEntity> findAllByCustomerId(String customerId);
 }
